// vue/src/store/generated/tendermint/bandchain/tendermint.bandchain.consuming/module/index.ts
// THIS FILE IS GENERATED AUTOMATICALLY. DO NOT MODIFY.

import { StdFee } from "@cosmjs/launchpad";
import { SigningStargateClient } from "@cosmjs/stargate";
import { Registry, OfflineSigner, EncodeObject, DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
import { Api } from "./rest";
import { MsgSendCreatePair } from "./types/consuming/tx";
import { MsgSendSellOrder } from "./types/consuming/tx";
import { MsgOracleData } from "./types/consuming/tx";


const types = [
  ["/tendermint.bandchain.consuming.MsgSendCreatePair", MsgSendCreatePair],
  ["/tendermint.bandchain.consuming.MsgSendSellOrder", MsgSendSellOrder],
  ["/tendermint.bandchain.consuming.MsgOracleData", MsgOracleData],
  
];
export const MissingWalletError = new Error("wallet is required");

const registry = new Registry(<any>types);

const defaultFee = {
  amount: [],
  gas: "200000",
};

interface TxClientOptions {
  addr: string
}

interface SignAndBroadcastOptions {
  fee: StdFee,
  memo?: string
}

const txClient = async (wallet: OfflineSigner, { addr: addr }: TxClientOptions = { addr: "http://localhost:26657" }) => {
  if (!wallet) throw MissingWalletError;

  const client = await SigningStargateClient.connectWithSigner(addr, wallet, { registry });
  const { address } = (await wallet.getAccounts())[0];

  return {
    signAndBroadcast: (msgs: EncodeObject[], { fee, memo }: SignAndBroadcastOptions = {fee: defaultFee, memo: ""}) => client.signAndBroadcast(address, msgs, fee,memo),
    msgSendCreatePair: (data: MsgSendCreatePair): EncodeObject => ({ typeUrl: "/tendermint.bandchain.consuming.MsgSendCreatePair", value: data }),
    msgSendSellOrder: (data: MsgSendSellOrder): EncodeObject => ({ typeUrl: "/tendermint.bandchain.consuming.MsgSendSellOrder", value: data }),
    msgOracleData: (data: MsgOracleData): EncodeObject => ({ typeUrl: "/tendermint.bandchain.consuming.MsgOracleData", value: data }),
    
  };
};

interface QueryClientOptions {
  addr: string
}

const queryClient = async ({ addr: addr }: QueryClientOptions = { addr: "http://localhost:1317" }) => {
  return new Api({ baseUrl: addr });
};

export {
  txClient,
  queryClient,
};
